bugfix/#4 correct readme and move args to own package
